requirements.txt
discord
python-decouple
pynacl
gTTS
pillow